[USER32] Implement IsServerSideWindow() (undocumented) (#9229)
CORE-19946
Adapted from PR #7577

Implement the undocumented `IsServerSideWindow()` function in user32.
It checks whether a window has a window procedure that resides in kernel
mode, and returns a boolean condition based on the appropriate internal
window state flag (`WNDS_SERVERSIDEWINDOWPROC`).

This function is not officially documented by Microsoft, but there exists
unofficial documentation[^1] for it.

It is required and used internally by uxtheme.dll from Windows XP/2003/Vista/7/etc.
The log spams with it a lot, when using uxtheme.dll from Windows in ReactOS.

[^1]: https://undoc.airesoft.co.uk/user32.dll/IsServerSideWindow.php